+++ extension/user/login.js
@@ -1,5 +1,5 @@
 const TokenHandler = require('../helpers/tokenHandler')
-const InvalidCredentialsError = require('../models/Errors/InvalidCredentialsError')
+const InvalidCredentials = require('../models/Errors/InvalidCredentials')
 
 /**
  * @typedef {Object} UserLoginInput
@@ -36,7 +36,7 @@ module.exports = function (context, input, cb) {
 
   login(th, userCredentials, strategy, (err, magentoTokenResponse) => {
     if (err) {
-      return cb(new InvalidCredentialsError('Invalid credentials were entered.'))
+      return cb(new InvalidCredentials('Invalid credentials were entered.'))
     }
 
     // delete token from device storage if it exists