+++ frontend/portals/GuestCheckoutLink/index.jsx
@@ -8,11 +8,11 @@ import styles from './style';
 import config from './config'
 
 /**
- * Check is the guest checkout disabled
+ * Check whether the guest checkout is disabled
  * @private
  * @returns {boolean}
  */
-const disableGuestCheckout = () => !config.getUserAccountSettings || config.getUserAccountSettings === 'required';
+const disableGuestCheckout = () => !config.getUserAccountSettings;
 
 /**
  * @returns {JSX}